chore: update colors file types with tests
